+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 7.6
 - PHP8 no xmlrpc ext: load xmlrpc compat functions in plugins/Moodle/client.php, xmlrpc.php, XML_RPC.php, XmlrpcDecoder.php & XmlrpcEncoder.php
 - Fix SQL Total points only select assignments for CP teacher (teacher may have changed) in Grades.php, InputFinalGrades.php, StudentGrades.php & GradebookBreakdown.php
 - Fix SQL Grades sort order in GradebookBreakdown.php
+- Add Login form link action hook in index.php & Actions.php
 
 Changes in 7.5
 --------------